Added confirm dialog

+++ src/app/modules/genre/pages/edit-genre/edit-genre.component.ts
@@ -50,8 +50,11 @@ export class EditGenreComponent implements OnInit {
 	}
 
 	deleteGenre() {
-		this.genreServ.deleteGenre(this.genre.id).subscribe(() => {
-			this.router.navigateByUrl('/genres');
+		this.translateServ.get('SHELL.MSGS.SURE').subscribe(res => {
+			if (!confirm(res)) return;
+			this.genreServ.deleteGenre(this.genre.id).subscribe(() => {
+				this.router.navigateByUrl('/genres');
+			});
 		});
 	}
 }
